refactor(Utils): move string case extensions to SharedKernel
