+++ DRN.Framework.EntityFramework/Attributes/DrnContextDefaultsAttribute.cs
@@ -1,5 +1,5 @@
+using DRN.Framework.SharedKernel.Extensions;
 using DRN.Framework.SharedKernel.Json;
-using DRN.Framework.Utils.Extensions;
 using DRN.Framework.Utils.Logging;
 using DRN.Framework.Utils.Settings;
 using Microsoft.EntityFrameworkCore;
